requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
reportlab
python-docx
jinja2
pytest
